pyproject.toml
[tool.poetry]
name = "software-engineering-for-internet-applications"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "software_engineering_for_internet_applications"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
turbo-django = "^0.3.0"
channels-redis = "^3.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
